Add files via upload
started off with trying to create the actual character that will be in the game : started with a single frame only 

make a function: initialisegraphics which takes a binary array and stores it into CGRAM:
created 8 rows of binary strings 5 numbers wide to represent the 5x8 sized pixels on the lcd 16x2 screen , 0 is where there is no light on and 1 is where the light is on. it is static so This frame is always there. 

the createchar operation loads this array (8x5 array/matrix) into the system RAM into index 0 (ie it is defined at the zero spot in the CGRAM/character generator ram.  

in the setup:
initialise the LCD with lcd begin by giving it the dimensions of the screen (16x2)

recall the initialise graphics function and it loads the binary array into CGRAM

indicate which block of the screen the character will be generated in using lcdcursor (0,0) is top left of the screen 

using lcd write, the custom character loaded in CGRAM at index 0 will be written on the screen . must use the syntax byte(0). this tell the program to specifically look for the thing that is in index 0 in the CGRAM , not the actual number 0 /ASCII 0